@@ -1,26 +1,21 @@
 name: mi_utem
-description: Plataforma académica para estudiantes de la Universidad Tecnológica
-  Metropolitana (UTEM)
+description: 'Plataforma académica para estudiantes de la Universidad Tecnológica Metropolitana (UTEM)'
 publish_to: none
 
-version: 2.11.2
+version: 3.0.0
 
 environment:
-  sdk: ">=2.14.4 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   awesome_notifications_core: ^0.8.1
-  awesome_notifications: any
-  awesome_notifications_fcm: any
+  awesome_notifications: ^0.8.2
+  awesome_notifications_fcm: ^0.8.0
   badges: ^2.0.2
   barcode_widget: ^2.0.3
   barcode_image: ^2.0.2
-  cached_network_image: ^3.2.0
   carousel_slider: ^4.0.0
-  charts_flutter:
-    git:
-      url: https://github.com/google/charts
-      path: charts_flutter
+  community_charts_flutter: ^1.0.2
   circular_profile_avatar: ^2.0.5
   clipboard: ^0.1.3
   cloud_firestore: ^4.3.1
@@ -29,8 +24,7 @@ dependencies:
   dio_cache_interceptor: ^3.2.6
   dio_cache_interceptor_hive_store: ^3.1.1
   dotted_border: ^2.0.0+2
-  extended_image: ^8.1.1
-  firebase_core: 2.11.0
+  firebase_core: 2.16.0
   firebase_analytics: 10.2.1
   firebase_in_app_messaging: ^0.7.0+10
   firebase_remote_config: ^3.0.9
@@ -38,7 +32,6 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_dotenv: ^5.0.2
-  flutter_keyboard_visibility: ^5.0.3
   flutter_markdown: ^0.6.1
   flutter_masked_text:
     git:
@@ -47,7 +40,6 @@ dependencies:
   flutter_spinkit: ^5.1.0
   flutter_windowmanager: ^0.2.0
   font_awesome_flutter: ^10.0.0
-  get: ^4.1.4
   gradient_widgets: ^0.6.0
   html: ^0.15.0
   image_editor: ^1.0.0
@@ -62,20 +54,30 @@ dependencies:
   qr_flutter: ^4.0.0
   recase: ^4.0.0
   responsive_framework: ^0.2.0
-  screenshot: ^2.1.0
-  sentry_flutter: ^7.10.1
+  sentry_flutter: ^7.12.0
   share_plus: ^7.2.1
   simple_gesture_detector: ^0.2.0
   url_launcher: ^6.1.7
   video_player: ^2.6.1
   dio_http_cache: ^0.3.0
   image: ^3.1.0
-  intl: ^0.17.0
+  intl: ^0.18.1
   get_storage: ^2.0.3
   hexcolor: ^3.0.1
   shared_preferences: ^2.0.20
   background_fetch: ^1.1.5
   flutter_uxcam: ^2.3.0
+  in_app_update: ^4.2.2
+  http: ^0.13.6
+  http_interceptor: ^1.0.2
+  screenshot: ^1.3.0
+  extended_image: ^7.0.2
+  cached_network_image: ^3.2.3
+  get_it: ^7.6.0
+  watch_it: ^1.0.2
+  extended_masked_text: ^2.3.1
+  crypto: ^3.0.3
+  listenable_collections: ^1.0.0
 
 dependency_overrides:
   qr: ^3.0.0
