@@ -1,3 +1,5 @@
 import 'package:logger/logger.dart';
 
-final logger = Logger();
+final logger = Logger(
+  printer: PrettyPrinter(),
+);
