@@ -1,28 +1,24 @@
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_uxcam/flutter_uxcam.dart';
-import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:mi_utem/config/constants.dart';
-import 'package:mi_utem/config/router.dart';
-import 'package:mi_utem/config/routes.dart';
-import 'package:mi_utem/controllers/calculator_controller.dart';
-import 'package:mi_utem/controllers/carreras_controller.dart';
-import 'package:mi_utem/services/analytics_service.dart';
-import 'package:mi_utem/services/auth_service.dart';
+import 'package:mi_utem/screens/splash_screen.dart';
 import 'package:mi_utem/services/background_service.dart';
 import 'package:mi_utem/services/notification_service.dart';
-import 'package:mi_utem/services/perfil_service.dart';
+import 'package:mi_utem/services_new/service_manager.dart';
 import 'package:mi_utem/themes/theme.dart';
 import 'package:responsive_framework/responsive_framework.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
 import 'services/remote_config/remote_config.dart';
+
+final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
+
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -34,6 +30,7 @@ void main() async {
   await RemoteConfigService.initialize();
   await NotificationService.initialize();
   await BackgroundService.initAndStart();
+  await registerServices();
   await SentryFlutter.init(
     (options) {
       options.dsn = Constants.sentryDsn;
@@ -52,37 +49,19 @@ class MiUtem extends StatefulWidget {
 class _MiUtemState extends State<MiUtem> {
   final FirebaseAnalytics analytics = FirebaseAnalytics.instance;
 
-  final calculatorController = Get.put(CalculatorController());
-
-  @override
-  void initState() {
-    if (AuthService.isLoggedIn()) {
-      final user = PerfilService.getLocalUsuario();
-      AnalyticsService.setUser(user);
-    } else {
-      AnalyticsService.removeUser();
-    }
-    super.initState();
-  }
-
   @override
   Widget build(BuildContext context) {
     FlutterUxcam.optIntoSchematicRecordings();
-    FlutterUxConfig config = FlutterUxConfig(
+    FlutterUxcam.startWithConfiguration(FlutterUxConfig(
       userAppKey: kDebugMode ? Constants.uxCamDevKey : Constants.uxCamProdKey,
       enableAutomaticScreenNameTagging: true,
       enableMultiSessionRecord: true,
-    );
-    FlutterUxcam.startWithConfiguration(config);
+    ));
 
-    return GetMaterialApp(
-      getPages: pages,
-      initialRoute: Routes.splash,
+    return MaterialApp(
+      home: SplashScreen(),
       debugShowCheckedModeBanner: false,
       title: 'Mi UTEM',
-      initialBinding: BindingsBuilder(() {
-        Get.put(CarrerasController(), permanent: true);
-      }),
       theme: MainTheme.theme,
       navigatorObservers: [
         FirebaseAnalyticsObserver(analytics: analytics),
