@@ -3,9 +3,8 @@
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:mi_utem/models/boletin_notas.dart';
-import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/services/boletin_service.dart';
-import 'package:mi_utem/services/horario_service.dart';
+import 'package:mi_utem/services/horario_repository.dart';
 import 'package:mi_utem/widgets/custom_expansion_tile.dart' as custom;
 import 'package:mi_utem/widgets/info_boletin_card.dart';
 import 'package:mi_utem/widgets/semestre_boletin_card.dart';
