@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:mi_utem/controllers/horario_controller.dart';
 import 'package:mi_utem/models/horario.dart';
+import 'package:mi_utem/services_new/interfaces/controllers/horario_controller.dart';
 import 'package:mi_utem/themes/theme.dart';
-import 'package:mi_utem/widgets/bloque_periodo_card.dart';
+import 'package:mi_utem/widgets/horario/bloque_periodo_card.dart';
+import 'package:watch_it/watch_it.dart';
 
 class HorarioPeriodsHeader extends StatelessWidget {
-  final HorarioController? controller;
   final Horario horario;
   final double periodHeight;
   final double width;
@@ -16,7 +16,6 @@ class HorarioPeriodsHeader extends StatelessWidget {
 
   const HorarioPeriodsHeader({
     Key? key,
-    this.controller,
     required this.horario,
     required this.periodHeight,
     required this.width,
@@ -37,8 +36,7 @@ class HorarioPeriodsHeader extends StatelessWidget {
                 inicio: horario.horasInicio[e.key],
                 intermedio: horario.horasIntermedio[e.key],
                 fin: horario.horasTermino[e.key],
-                active: showActivePeriod &&
-                    controller?.indexOfCurrentPeriod == e.key,
+                active: showActivePeriod && di.get<HorarioController>().indexOfCurrentPeriod == e.key,
                 height: periodHeight,
                 width: width,
                 backgroundColor: backgroundColor,
