@@ -3,7 +3,6 @@ import 'dart:developer';
 import 'package:dio/dio.dart';
 import 'package:dio_http_cache/dio_http_cache.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:mi_utem/services/auth_service.dart';
 
 class DioMiUtemClient {
   static const bool isProduction = bool.fromEnvironment('dart.vm.product');
@@ -45,7 +44,7 @@ class AuthInterceptor extends QueuedInterceptor {
     final RequestInterceptorHandler handler,
   ) async {
     try {
-      final token = AuthService.getToken();
+      final token = "";
 
       options._setAuthenticationHeader(token);
 
@@ -75,7 +74,7 @@ class AuthInterceptor extends QueuedInterceptor {
 
     // Force refresh auth token
     try {
-      final token = await AuthService.refreshToken();
+      final token = "";
 
       log("Refreshing token, attempt $attempt...");
 
@@ -93,7 +92,7 @@ class AuthInterceptor extends QueuedInterceptor {
   }
 
   Future<void> _onErrorRefreshingToken() async {
-    AuthService.invalidateToken();
+
   }
 }
 
