@@ -6,40 +6,37 @@ class CustomErrorWidget extends StatelessWidget {
   final String title;
   final Object? error;
 
-  CustomErrorWidget({
-    Key? key,
+  const CustomErrorWidget({
+    super.key,
     this.emoji = "😕",
     this.title = "Ocurrió un error inesperado",
     this.error,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.all(20),
+    return Padding(
+      padding: const EdgeInsets.all(20),
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          Text(
-            emoji,
+          Text(emoji,
             textAlign: TextAlign.center,
-            style: TextStyle(
+            style: const TextStyle(
               fontSize: 50,
             ),
           ),
-          Container(height: 15),
-          Text(
-            title,
+          const SizedBox(height: 15),
+          Text(title,
             textAlign: TextAlign.center,
-            style: TextStyle(
+            style: const TextStyle(
               fontWeight: FontWeight.bold,
               fontSize: 18,
             ),
           ),
           if (error != null) ...[
-            Container(height: 15),
-            Text(
-              error.toString(),
+            const SizedBox(height: 15),
+            Text("$error",
               textAlign: TextAlign.center,
             ),
           ],
